Cleaned up object selection + Gamepad movement fix
Moved all of the selection code into playerClass since object selection is purely a local function specific to the player and does not need to be in the global animate function.
Also the gamepad movement was previously a bit buggy due to direction only taking into account the playerGroup direction, and not the nested VR camera direction.